# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FIRS E-Invoicing Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  fullName: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  fullName: string;
  onboardingComplete: boolean;
  subscriptionPlan?: string | null;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface BusinessProfileRequest {
  companyName: string;
  tin: string;
  registeredAddress: string;
  city?: string;
  state?: string;
  contactEmail: string;
  contactPhone: string;
}

export interface BusinessProfile {
  id: number;
  userId: number;
  companyName: string;
  tin: string;
  registeredAddress: string;
  city?: string | null;
  state?: string | null;
  contactEmail: string;
  contactPhone: string;
  firsConnected: boolean;
  firsApiKey?: string | null;
  subscriptionPlan?: string | null;
  createdAt?: string;
}

export interface FirsConnectRequest {
  apiKey: string;
  taxOfficeCode: string;
}

export interface FirsConnectResponse {
  connected: boolean;
  message: string;
  taxOfficeName?: string | null;
}

export type SubscriptionRequestPlan =
  (typeof SubscriptionRequestPlan)[keyof typeof SubscriptionRequestPlan];

export const SubscriptionRequestPlan = {
  basic: "basic",
  enterprise: "enterprise",
} as const;

export interface SubscriptionRequest {
  plan: SubscriptionRequestPlan;
}

export interface SubscriptionResponse {
  plan: string;
  activatedAt: string;
  message: string;
}

export interface CustomerRequest {
  name: string;
  tin: string;
  email?: string | null;
  phone?: string | null;
  address: string;
  city?: string | null;
  state?: string | null;
}

export interface Customer {
  id: number;
  userId: number;
  name: string;
  tin: string;
  email?: string | null;
  phone?: string | null;
  address: string;
  city?: string | null;
  state?: string | null;
  createdAt?: string;
}

export interface TinLookupResponse {
  found: boolean;
  tin: string;
  name?: string | null;
  address?: string | null;
  registrationStatus?: string | null;
}

export interface LineItem {
  description: string;
  quantity: number;
  unitPrice: number;
  vatRate: number;
  amount: number;
}

export interface InvoiceRequest {
  customerId: number;
  invoiceNumber: string;
  invoiceDate: string;
  dueDate?: string | null;
  lineItems: LineItem[];
  subtotal: number;
  vatAmount: number;
  totalAmount: number;
  currency?: string;
  notes?: string | null;
}

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus];

export const InvoiceStatus = {
  draft: "draft",
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface Invoice {
  id: number;
  userId: number;
  customerId: number;
  customer?: Customer | null;
  invoiceNumber: string;
  invoiceDate: string;
  dueDate?: string | null;
  lineItems: LineItem[];
  subtotal: number;
  vatAmount: number;
  totalAmount: number;
  currency: string;
  status: InvoiceStatus;
  firsReferenceId?: string | null;
  rejectionReasons?: string[] | null;
  notes?: string | null;
  submittedAt?: string | null;
  createdAt?: string;
  updatedAt?: string;
}

export interface ValidationError {
  field: string;
  message: string;
  code: string;
}

export interface ValidationResponse {
  valid: boolean;
  errors: ValidationError[];
}

export type SubmissionResponseStatus =
  (typeof SubmissionResponseStatus)[keyof typeof SubmissionResponseStatus];

export const SubmissionResponseStatus = {
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface SubmissionResponse {
  success: boolean;
  status: SubmissionResponseStatus;
  firsReferenceId?: string | null;
  message: string;
  rejectionReasons?: string[] | null;
  submittedAt: string;
}