# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FIRS E-Invoicing Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /business/profile:
    get:
      operationId: getBusinessProfile
      tags: [business]
      summary: Get business profile
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
        "404":
          description: No profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createOrUpdateBusinessProfile
      tags: [business]
      summary: Create or update business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessProfileRequest"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
  /business/firs-connect:
    post:
      operationId: connectFirs
      tags: [business]
      summary: Connect to FIRS API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FirsConnectRequest"
      responses:
        "200":
          description: FIRS connection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FirsConnectResponse"
  /business/subscription:
    post:
      operationId: selectSubscription
      tags: [business]
      summary: Select a subscription plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionRequest"
      responses:
        "200":
          description: Subscription activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionResponse"
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequest"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /customers/tin/{tin}:
    get:
      operationId: lookupCustomerByTin
      tags: [customers]
      summary: Look up customer by TIN (simulates FIRS lookup)
      parameters:
        - name: tin
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer data from TIN lookup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TinLookupResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        fullName:
          type: string
      required:
        - email
        - password
        - fullName
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        onboardingComplete:
          type: boolean
        subscriptionPlan:
          type: string
          nullable: true
      required:
        - id
        - email
        - fullName
        - onboardingComplete
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    BusinessProfileRequest:
      type: object
      properties:
        companyName:
          type: string
        tin:
          type: string
        registeredAddress:
          type: string
        city:
          type: string
        state:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
      required:
        - companyName
        - tin
        - registeredAddress
        - contactEmail
        - contactPhone
    BusinessProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        companyName:
          type: string
        tin:
          type: string
        registeredAddress:
          type: string
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        contactEmail:
          type: string
        contactPhone:
          type: string
        firsConnected:
          type: boolean
        firsApiKey:
          type: string
          nullable: true
        subscriptionPlan:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - companyName
        - tin
        - registeredAddress
        - contactEmail
        - contactPhone
        - firsConnected
    FirsConnectRequest:
      type: object
      properties:
        apiKey:
          type: string
        taxOfficeCode:
          type: string
      required:
        - apiKey
        - taxOfficeCode
    FirsConnectResponse:
      type: object
      properties:
        connected:
          type: boolean
        message:
          type: string
        taxOfficeName:
          type: string
          nullable: true
      required:
        - connected
        - message
    SubscriptionRequest:
      type: object
      properties:
        plan:
          type: string
          enum: [basic, enterprise]
      required:
        - plan
    SubscriptionResponse:
      type: object
      properties:
        plan:
          type: string
        activatedAt:
          type: string
          format: date-time
        message:
          type: string
      required:
        - plan
        - activatedAt
        - message
    CustomerRequest:
      type: object
      properties:
        name:
          type: string
        tin:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
      required:
        - name
        - tin
        - address
    Customer:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        tin:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - tin
        - address
    TinLookupResponse:
      type: object
      properties:
        found:
          type: boolean
        tin:
          type: string
        name:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        registrationStatus:
          type: string
          nullable: true
      required:
        - found
        - tin
...