# lib/api-zod/src/generated/types/subscriptionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FIRS E-Invoicing Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionRequestPlan } from "./subscriptionRequestPlan";

export interface SubscriptionRequest {
  plan: SubscriptionRequestPlan;
}